+++ src/schemas/account_types/banking/bnpl.py
@@ -164,6 +164,7 @@ class BNPLAccountBase(AccountBase):
             "PayPal Pay in 4",
             "Uplift",
             "Quadpay",
+            "SplitIt",
             "Other",
         ]